[refactor] 고객센터 패키지명 통합 리팩토링
